EventCounterQueue(): renamed
