+++ src/queutils/__init__.py
@@ -2,15 +2,15 @@ from .countable import Countable as Countable
 from .asyncqueue import AsyncQueue as AsyncQueue
 from .iterablequeue import IterableQueue as IterableQueue, QueueDone as QueueDone
 from .filequeue import FileQueue as FileQueue
-from .categorycounterqueue import (
+from .eventcounterqueue import (
     QCounter as QCounter,
-    CategoryCounterQueue as CategoryCounterQueue,
+    EventCounterQueue as EventCounterQueue,
 )
 
 __all__ = [
     "asyncqueue",
     "countable",
-    "categorycounterqueue",
+    "eventcounterqueue",
     "filequeue",
     "iterablequeue",
 ]
